+++ projects/pastanaga/src/lib/dropdown/dropdown-separator.component.ts
@@ -3,6 +3,5 @@ import { Component } from '@angular/core';
 @Component({
     selector: 'pa-dropdown-separator',
     template: '<li class="pa-dropdown-separator" role="separator"></li>',
-    styleUrls: ['./_dropdowns.scss'],
 })
 export class DropdownSeparatorComponent {}